Helper function to create the parameters for netsim_run_swfcalib_scenario

Helper function to create the parameters for netsim_run_swfcalib_scenario

Usage

netsim_swfcalib_output_setup(
  path_to_x,
  param,
  init,
  control,
  calib_object,
  n_rep,
  n_cores,
  output_dir,
  libraries,
  save_pattern
)

Arguments

path_to_x

Path to a Fitted network model object saved with saveRDS. (See the x argument to the EpiModel::netsim function)

param

Model parameters, as an object of class param.net.

init

Initial conditions, as an object of class init.net.

control

Control settings, as an object of class control.net.

calib_object

a formatted calibration object

n_rep

The number of replication to be run for each scenario.

n_cores

number of cores to run the processing on

output_dir

The folder where the simulation files are to be stored.

libraries

A character vector containing the name of the libraries required for the model to run. (e.g. EpiModelHIV or EpiModelCOVID)

save_pattern

A character vector of what should be kept in the final netsim objects. It can contain the names of the elements as well as: "simple" (defautlt) to only keep "epi", "param" and "control"; "restart" to get the elements required to restart from such file; "all" to not trim the object at all. c("simple", "el.cuml") is an example of a valid pattern to save "epi", "param", "control" and "el.cuml". If control$raw.output is TRUE, this parameter has no effect and the full result is saved.