EpiModel Extensions for High-Performance Computing

Package:	EpiModelHPC

Type:	Package

Version:	2.2.0

Date:	2022-11-29

License:	GPL-3

LazyLoad:	yes

Details

EpiModel provides tools for the mathematical modeling of infectious diseases. Supported model classes include stochastic network models, which rely on the statistical framework of exponential-family random graph models (ERGMs) that evolve over time. This allows for modeling of disease-related contacts with duration, such as ongoing sexual partnerships.

The level of statistical complexity of these models, based in Markov-chain Monte Carlo (MCMC) simulation, results in computationally intensive simulation processes. The goal of EpiModelHPC is to provide a standardized framework for extending EpiModel to run on modern high-performance computing (HPC) systems.

References

The main website for EpiModel is at http://epimodel.org/. The source code for this extension package is hosted on Github at http://github.com/statnet/EpiModelHPC. Bug reports and feature requests may be filed there.