Run one netsim call with a scenario and saves the results deterministically

This inner function is called by netsim_scenarios and step_tmpl_netsim_scenarios.

Usage

netsim_run_one_scenario(
  scenario,
  batch_num,
  path_to_x,
  param,
  init,
  control,
  libraries,
  output_dir,
  n_batch,
  n_rep,
  n_cores,
  save_all,
  save_elements
)

Arguments

scenario

A single "EpiModel scenario" to be used in the simulation

batch_num

The batch number, calculated from the number of replications and CPUs required.

path_to_x

Path to a Fitted network model object saved with saveRDS. (See the x argument to the EpiModel::netsim function)

param

Model parameters, as an object of class param.net.

init

Initial conditions, as an object of class init.net.

control

Control settings, as an object of class control.net.

libraries

A character vector containing the name of the libraries required for the model to run. (e.g. EpiModelHIV or EpiModelCOVID)

output_dir

The folder where the simulation files are to be stored.

n_batch

The number of batches to be run ceiling(n_rep / n_cores).

n_rep

The number of replication to be run for each scenario.

n_cores

The number of CPUs on which the simulations will be run.

save_all

A flag instructing to save the result of the EpiModel::netsim call as is if TRUE.

save_elements

A character vector of elements to keep from the netsim object if save_all is FALSE

Checkpointing

This function takes care of editing .checkpoint.dir to create unique sub directories for each scenario. The EpiModel::control.net way of setting up checkpoints can be used transparently.